requirements.txt
nltk
spacy==2.3.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
pyresparser
numpy==1.26.4
openai
rapidfuzz
pymupdf
python-docx
pyrate-limiter==2.4.7